Standardize OTEL error logging format to match application logs
